modified hollow crystal textures

+++ build.sh
@@ -1,11 +1,12 @@
 #!/bin/bash -e
 
 usage_exit() {
-  echo "Usage $0 [-h] [-b, --behavior] [-s, --src] [-r, --resource]" 1>&2
+  echo "Usage $0 [-h] [-b, --behavior] [-d, --dist] [-s, --src] [-r, --resource]" 1>&2
   echo
   echo optional arguments:
   echo "  -h, --help            show this help message and exit"
   echo "  -b, --behavior        build behavior pack"
+  echo "  -d, --dist            only make dist directory"
   echo "  -s, --s               build ts source"
   echo "  -r, --resource        build resource pack"
   exit 1
@@ -16,13 +17,15 @@ BEHAVIOR=
 SOURCE=
 RESOURCE=
 
-if [ "$1" = "" ]; then
+echo $1
+
+if [ "$1" = "" -o "$1" = "-d" -o "$1" = "--dist" ]; then
   BEHAVIOR=1
   SOURCE=1
   RESOURCE=1
 fi
 
-OPTIONS=`getopt -o hbsr -l help,behavior,source,resource -- "$@"`
+OPTIONS=`getopt -o hbdsr -l help,behavior,dist,source,resource -- "$@"`
 eval set -- "$OPTIONS"
 while true
 do
@@ -32,6 +35,10 @@ do
       ;;
     -b|--behavior)
       BEHAVIOR=1
+      shift
+      ;;
+    -d|--dist)
+      DIST=1
       shift
       ;;
     -s|--source)
@@ -53,29 +60,39 @@ done
 if [ "$BEHAVIOR" = 1 -a "$SOURCE" = 1 -a "$RESOURCE" = 1 ]; then
   rm -rf dist/*
   mkdir dist/nacht_server_BP dist/nacht_server_RP
-  rm -rf ../development_behavior_packs/nacht_server_BP ../development_resource_packs/nacht_server_RP
-  mkdir ../development_behavior_packs/nacht_server_BP ../development_resource_packs/nacht_server_RP
+  if [ "$DIST" != "1" ]; then
+    rm -rf ../development_behavior_packs/nacht_server_BP ../development_resource_packs/nacht_server_RP
+    mkdir ../development_behavior_packs/nacht_server_BP ../development_resource_packs/nacht_server_RP
+  fi
 elif [ "$BEHAVIOR" = 1 -a "$SOURCE" = 1 ]; then
   rm -rf dist/nacht_server_BP
   mkdir dist/nacht_server_BP
-  rm -rf ../development_behavior_packs/nacht_server_BP
-  mkdir ../development_behavior_packs/nacht_server_BP
+  if [ "$DIST" != "1" ]; then
+    rm -rf ../development_behavior_packs/nacht_server_BP
+    mkdir ../development_behavior_packs/nacht_server_BP
+  fi
 else
   if [ "$BEHAVIOR" = 1 ]; then
     for d in `ls dist/nacht_server_BP`; do
       if [ -d "dist/nacht_server_BP/$d" -a "$d" != "scripts" ]; then
         rm -rf dist/nacht_server_BP/$d
-        rm -rf ../development_behavior_packs/nacht_server_BP/$d
+        if [ "$DIST" != "1" ]; then
+          rm -rf ../development_behavior_packs/nacht_server_BP/$d
+        fi
       fi
     done
   fi
   if [ "$SOURCE" = 1 ]; then
     rm -rf dist/nacht_server_BP/scripts
-    rm -rf ../development_behavior_packs/nacht_server_BP/scripts
+    if [ "$DIST" != "1" ]; then
+      rm -rf ../development_behavior_packs/nacht_server_BP/scripts
+    fi
   fi
   if [ "$RESOURCE" = 1 ]; then
     rm -rf dist/nacht_server_RP
-    rm -rf ../development_resource_packs/nacht_server_RP
+    if [ "$DIST" != "1" ]; then
+      rm -rf ../development_resource_packs/nacht_server_RP
+    fi
   fi
 fi
 
@@ -87,7 +104,9 @@ if [ "$SOURCE" = 1 ]; then
   cp node_modules/@minecraft/vanilla-data/lib/index.js nacht_server_BP/scripts/types/
 
   cp -r nacht_server_BP/scripts dist/nacht_server_BP/
-  cp -r dist/nacht_server_BP ../development_behavior_packs
+  if [ "$DIST" != "1" ]; then
+    cp -r dist/nacht_server_BP ../development_behavior_packs
+  fi
 fi
 
 if [ "$BEHAVIOR" = 1 ]; then
@@ -97,16 +116,22 @@ if [ "$BEHAVIOR" = 1 ]; then
     fi
   done
 
-  cp -r dist/nacht_server_BP ../development_behavior_packs
+  if [ "$DIST" != "1" ]; then
+    cp -r dist/nacht_server_BP ../development_behavior_packs
+  fi
 fi
 
 if [ "$SOURCE" = 1 -o "$BEHAVIOR" = 1 ]; then
   cp nacht_server_BP/manifest.json nacht_server_BP/pack_icon.png dist/nacht_server_BP/
-  cp -r dist/nacht_server_BP ../development_behavior_packs
+  if [ "$DIST" != "1" ]; then
+    cp -r dist/nacht_server_BP ../development_behavior_packs
+  fi
 fi
 
 if [ "$RESOURCE" = 1 ]; then
   cp -r nacht_server_RP dist/
 
-  cp -r dist/nacht_server_RP ../development_resource_packs
+  if [ "$DIST" != "1" ]; then
+    cp -r dist/nacht_server_RP ../development_resource_packs
+  fi
 fi
